Custom loading Animation for your MC Eternal Modpack(Using just Forge Minecraft Loader so no extra mods needed!).
I created a custom loading Animation for your MC Eternal Modpack(Using just Forge Minecraft Loader so no extra mods needed!).
Other Modpack developer had also his own little something that replace the Anvil with hammer thing. So it can be that hard right? I was so wrong… There was near to zero information on the internet and no clarification how big the file should be or what formats are supported.
This processed was very frustrating and I have no idea why this file now works and why some other not (even when all have the same export settings) I spent around 10h to get this in something “solid”

It is not the best-looking thing I know, but there were many limitations what forge accepted without crashing mc. So, it is a little bit minimalistic for some volks.
I tried to hurry because the Update 1.6.1 is just around the corner to eventually made an official way into your modpack. ❤